celllist

+++ mindone/transformers/models/gpt2/modeling_gpt2.py
@@ -201,7 +201,7 @@ class GPT2Attention(nn.Cell):
         attn_weights = mint.softmax(attn_weights, dim=-1)
 
         # Downcast (if necessary) back to V's dtype (if in mixed-precision) -- No-Op otherwise
-        attn_weights = attn_weights.type(value.dtype)
+        attn_weights = attn_weights.to(value.dtype)
         attn_weights = self.attn_dropout(attn_weights)
 
         # Mask heads if we want to
